+++ src/main/java/com/caorderapi/dto/ProductCacheDto.java
@@ -4,10 +4,10 @@ import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductCacheDto(
-         UUID productId,
-         String sku,
-         BigDecimal price,
-         Integer stockQuantity
+        UUID productId,
+        String sku,
+        BigDecimal price,
+        Long stockQuantity
 ) {
 }
